opendal_service_oss

Struct Oss

pub struct Oss { /* private fields */ }

Aliyun Object Storage Service (OSS) support

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::OssConfig for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Example

Via Builder

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_oss::Oss;

#[tokio::main]
async fn main() -> Result<()> {
    // Create OSS backend builder.
    let mut builder = Oss::default()
        // Set the root for oss, all operations will happen under this root.
        //
        // NOTE: the root must be absolute path.
        .root("/path/to/dir")
        // Set the bucket name, this is required.
        .bucket("test")
        // Set the endpoint.
        //
        // For example:
        // - "https://oss-ap-northeast-1.aliyuncs.com"
        // - "https://oss-hangzhou.aliyuncs.com"
        .endpoint("https://oss-cn-beijing.aliyuncs.com")
        // Set the access_key_id and access_key_secret.
        //
        // OpenDAL will try load credential from the env.
        // If credential not set and no valid credential in env, OpenDAL will
        // send request without signing like anonymous user.
        .access_key_id("access_key_id")
        .access_key_secret("access_key_secret");

    let op: Operator = Operator::new(builder)?;

    Ok(())
}

Implementations

impl OssBuilder

pub fn root(self, root: &str) -> Self

Set root of this backend.

All operations will happen under this root.

pub fn bucket(self, bucket: &str) -> Self

Set bucket name of this backend.

pub fn endpoint(self, endpoint: &str) -> Self

Set endpoint of this backend.

pub fn addressing_style(self, addressing_style: &str) -> Self

Set addressing style for the endpoint.

Available values: virtual, cname, path.

virtual: Use virtual addressing style, i.e. http://bucket.oss-<region>.aliyuncs.com/object

cname: Use cname addressing style, i.e. http://mydomain.com/object with mydomain.com bound to your bucket.

path: Use path addressing style. i.e. http://oss-<region>.aliyuncs.com/bucket/object

If not set, default value is virtual.

pub fn enable_versioning(self, _enabled: bool) -> Self

👎Deprecated since 0.57.0:
OSS versioning capability is enabled by default and this option is no longer needed.

Deprecated: OSS versioning capability is enabled by default.

pub fn presign_endpoint(self, endpoint: &str) -> Self

Set an endpoint for generating presigned urls.

You can offer a public endpoint like https://oss-cn-beijing.aliyuncs.com to return a presinged url for public accessors, along with an internal endpoint like https://oss-cn-beijing-internal.aliyuncs.com to access objects in a faster path.

If presign_endpoint is set, we will use presign_endpoint on generating presigned urls.

if not, we will use endpoint as default.

pub fn presign_addressing_style(self, addressing_style: &str) -> Self

Set addressing style for presign endpoint.

Similar to setting addressing style for endpoint.

If both presign_endpoint and presign_addressing_style are not set, they are the same as endpoint’s configurations.

If presign_endpoint is set, but presign_addressing_style is not set, default value is virtual.

pub fn access_key_id(self, v: &str) -> Self

Set access_key_id of this backend.

If access_key_id is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn access_key_secret(self, v: &str) -> Self

Set access_key_secret of this backend.

If access_key_secret is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn security_token(self, security_token: &str) -> Self

Set security_token for this backend.

If security_token is set, we will take user’s input first.

If not, we will try to load it from environment.

pub fn server_side_encryption(self, v: &str) -> Self

Set server_side_encryption for this backend.

Available values: AES256, KMS.

Reference: https://www.alibabacloud.com/help/en/object-storage-service/latest/server-side-encryption-5 Brief explanation: There are two server-side encryption methods available: SSE-AES256: 1. Configure the bucket encryption mode as OSS-managed and specify the encryption algorithm as AES256. 2. Include the x-oss-server-side-encryption parameter in the request and set its value to AES256. SSE-KMS: 1. To use this service, you need to first enable KMS. 2. Configure the bucket encryption mode as KMS, and specify the specific CMK ID for BYOK (Bring Your Own Key) or not specify the specific CMK ID for OSS-managed KMS key. 3. Include the x-oss-server-side-encryption parameter in the request and set its value to KMS. 4. If a specific CMK ID is specified, include the x-oss-server-side-encryption-key-id parameter in the request, and set its value to the specified CMK ID.

pub fn server_side_encryption_key_id(self, v: &str) -> Self

Set server_side_encryption_key_id for this backend.

Notes

This option only takes effect when server_side_encryption equals to KMS.

pub fn batch_max_operations(self, _delete_max_size: usize) -> Self

👎Deprecated since 0.57.0:
OSS delete batch capability is enabled by default and this option is no longer needed.

Deprecated: OSS delete batch capability is enabled by default.

pub fn delete_max_size(self, _delete_max_size: usize) -> Self

👎Deprecated since 0.57.0:
OSS delete batch capability is enabled by default and this option is no longer needed.

Deprecated: OSS delete batch capability is enabled by default.

pub fn skip_signature(self) -> Self

Skip signature will skip loading credentials and signing requests.

pub fn allow_anonymous(self) -> Self

👎Deprecated since 0.57.0:
Please use skip_signature instead of allow_anonymous

Allow anonymous will allow opendal to send request without signing when credential is not loaded.

pub fn role_arn(self, role_arn: &str) -> Self

Set role_arn for this backend.

If role_arn is set, we will use already known config as source credential to assume role with role_arn.

pub fn role_session_name(self, role_session_name: &str) -> Self

Set role_session_name for this backend.

pub fn oidc_provider_arn(self, oidc_provider_arn: &str) -> Self

Set oidc_provider_arn for this backend.

pub fn oidc_token_file(self, oidc_token_file: &str) -> Self

Set oidc_token_file for this backend.

pub fn sts_endpoint(self, sts_endpoint: &str) -> Self

Set sts_endpoint for this backend.

pub fn external_id(self, v: &str) -> Self

Set external_id for this backend.